# perf: Speed up maketx --envlatl when multithreaded by over 10x.